Messing around with colors

+++ Microsoft.PowerShell_profile.ps1
@@ -3,8 +3,11 @@ Import-Module PSUrl
 Import-Module Aliases
 Import-Module SyncMeUp
 
-(Get-Content ~\.last) | set-location
-rm ~\.last
+if(Test-Path ~\.last) {
+    (Get-Content ~\.last) | set-location
+    rm ~\.last
+}
+
 
 
 # Makes git diff work
